Move fundref function to funding module.
